Actualizacion de productos lenta, falta optimizar

+++ database/migrations/2020_09_08_184220_create_celler_section_table.php
@@ -16,7 +16,7 @@ class CreateCellerSectionTable extends Migration{
             $table->string('name', 25);
             $table->string('alias', 20);
             $table->string('path', 40);
-            $table->tinyInteger('root');
+            $table->mediumInteger('root');
             $table->tinyInteger('deep');
             $table->json('details');
             $table->unsignedMediumInteger('_celler');
